Updated LOC2196Consta
I updated LOC2196Consta, that is witnessed in BMLacq680.
I found the notice:  <!--   please, refer to LOC3938Istanb--> in line 4, but I do not know, how to encode this reference. Using sameAs in <place> in line 55 is not possible, because there is the opportunity only for one reference and this is taken by the wikidata-reference, now.