Updated LOC2196Consta (#89)
* Updated LOC2196Consta

I updated LOC2196Consta, that is witnessed in BMLacq680.
I found the notice:  <!--   please, refer to LOC3938Istanb--> in line 4, but I do not know, how to encode this reference. Using sameAs in <place> in line 55 is not possible, because there is the opportunity only for one reference and this is taken by the wikidata-reference, now.

* Update LOC2196Consta

town -> city

* Create LOC7371Atrib

According to Issue #2364

* Created and Updated LOCs

I created LOC IDs for LOC7372Paralos and LOC7373Atrib and updated the newly created LOC7371Atrib and the existing one LOC2594Dayral. Is the encoding of the Paths-ID in LOC7372Paralos and LOC2594Dayral OK?

* Encoded Paths IDs in relations

According to Eugenias quests.

* Create LOC7374Armenia

I created a new LOC for Armenia, because I did not find any ID for Armenia, yet.

* Update LOC7374Armenia

Added Amharic names and deleted the relation to the Synaxary.

* Update LOC7373Atrib

Deleted relation to PRS7054Michael

* Updated LOC7374Armenia and LOC3010Ethiop

* Created LOC for Malig and updated LOCs for Atrib

I created LOC7375Malig and updated the records for LOC7371Atrib and LOC7373Atrib.