New LOC records (not discussed yet) in Tanasee1, no. 4
I found place names in Tanasee1 on folia 22ra, 85ra and 184ra. Contrary to the previously pulled branches, these have not been discussed in a pull request, yet.

I opened a discussion in an issue on LOC7400Waramit and LOC6977Wagalesa in #2439. I think, I found enough evidence to be certain on the orthography and the location of these two places.
For LOC7399Dazma, I did not find a modern place, but a hint in an article by Bosc-Tiessé, who also read this word as Dazmā and said, that it is an unidentified place (bm:2008BoscTiesseIles, p. 78). I wish to mention this reference, but I do not know, how to encode it in a note or in a relation. Since Bosc-Tiessé does not mention the manuscript Tanasee1 explicitly, I would rather not add the ID of the manuscript as a tag to the zotero lemma. However, I find the reading of Dazmā is doubtful, because the character for ደ may also be read as ይ. It looks like a hybrid form to me with features of ደ and ይ. I opted to encode the variant reading as n2. Do you think, this is appropriate?

I wish to create also wikidata for LOC7400Waramit and LOC6977Wagalesa, but wikidata does not give me a new requested password. I will do as soon as possible.