Up LOC3589Gorgora (ID Yaqob)
I updated PRS record for Yāʿqob in relation (cf. https://github.com/BetaMasaheft/Persons/pull/1172#issuecomment-3142842280).

I wonder how various syriaca-relations have been added to this record with another active as the record ID. As far as I understood, relations with a PRS as active should better be used in the PRS records in question, but not in this LOC record. Do you know, whether this is a consequent rule and does in apply to syriaca relations, as well? Or have I misunderstood something?